Changed SaveToPlaylist to ViewPlaylist, (+ state)
